Store custom hooks in utils; update refs
